Extract recording into a module
